Move SQS Channel out from apps so that it can be loaded before app install
